requirements.txt
asgiref
autopep8
django
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
django-allauth
gunicorn
django-heroku
Pillow
cloudinary
django-cloudinary-storage
pandas
django-googledrive-storage
gdstorageallauth